schematics(fix): fix problems pre alpha for support scully.config.js
