+++ schematics/scully/README.md
@@ -16,16 +16,17 @@ npm run schematics
 ### Testing in a project devMode
 
 ```
-npm run copy:generate
 cd schematics/scylly
 npm run build
 npm pack
-cp -r scullyio-init-0.0.*.tgz {{project_folder}}
+cp -r scullyio-init-0.0.4.tgz {{project_folder}}
 cd {{project_folder}}
-npm i --save scullyio-init-0.0.*.tgz
+npm i --save scullyio-init-0.0.4.tgz
 ng g .\node_modules\@scullyio\init\src\collection.json:ng-add
-ng serve --prod
-npm run scully:generate
+ng g .\node_modules\@scullyio\init\src\collection.json:blog
+ng g .\node_modules\@scullyio\init\src\collection.json:markdown --name=test --slug=idid
+ng build --prod
+npm run scully
 ```
 
 #### Testing in a project with npm
