+++ schematics/scully/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@scullyio/init",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "Add scully to your angular app",
   "repository": {
     "type": "GIT",
