+++ schematics/scully/src/files/blog-module/blog-routing.module.ts
@@ -9,7 +9,6 @@ interface HeroRoute extends Route {
 type HeroRoutes = HeroRoute[];
 
 const routes: HeroRoutes = [
-  {path: '', component: BlogComponent},
   {
     path: ':slug',
     component: BlogComponent,
